+++ src/test/java/ru/netology/service/SalesServiceTest.java
@@ -13,10 +13,10 @@ public class SalesServiceTest {
     public void shouldFindSumSales() {
         SalesService service = new SalesService();
 
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int expectedSumSales = 180;
-        int actualSumSales = service.getSumSales(sales);
+        long expectedSumSales = 180;
+        long actualSumSales = service.getSumSales(sales);
 
         Assertions.assertEquals(expectedSumSales, actualSumSales);
     }
@@ -25,10 +25,10 @@ public class SalesServiceTest {
     public void shouldFindAverageSales() {
         SalesService service = new SalesService();
 
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int expectedAverageSales = 15;
-        int actualAverage = service.getAverageSales(sales);
+        long expectedAverageSales = 15;
+        long actualAverage = service.getAverageSales(sales);
 
         Assertions.assertEquals(expectedAverageSales, actualAverage);
     }
@@ -37,7 +37,7 @@ public class SalesServiceTest {
     public void shouldFindMaxMonthSales() {
         SalesService service = new SalesService();
 
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expectedMaxMonth = 8;
         int actualMaxMonth = service.getMaxMonth(sales);
@@ -49,7 +49,7 @@ public class SalesServiceTest {
     public void shouldFindMinMonthSales() {
         SalesService service = new SalesService();
 
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expectedMinMonth = 9;
         int actualMinMonth = service.getMinMonth(sales);
@@ -61,7 +61,7 @@ public class SalesServiceTest {
     public void shouldCountMonthSalesBelowAverage() {
         SalesService service = new SalesService();
 
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedCntMonthBelowAverage = 5;
         int actualCntMonthBelowAverage = service.getCountMonthSalesBelowAverage(sales);
 
@@ -73,7 +73,7 @@ public class SalesServiceTest {
     public void shouldCountMonthSalesAboveAverage() {
         SalesService service = new SalesService();
 
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedCntMonthAboveAverage = 5;
         int actualCntMonthAboveAverage = service.getCountMonthSalesAboveAverage(sales);
